web/pyproject.toml
[tool.poetry]
name = "nfl-rushing"
version = "0.1.0"
description = ""
authors = ["Brandon Bil <me@brandonbil.xyz>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
